refactor: extract hooks to their own directory
