+++ src/app/Scenes/HomeView/Sections/HomeViewSectionCardsChips.tsx
@@ -6,7 +6,7 @@ import { SectionTitle } from "app/Components/SectionTitle"
 import { getSnapToOffsets } from "app/Scenes/CollectionsByCategory/CollectionsChips"
 import { HomeViewSectionSentinel } from "app/Scenes/HomeView/Components/HomeViewSectionSentinel"
 import { SectionSharedProps } from "app/Scenes/HomeView/Sections/Section"
-import { useHomeViewTracking } from "app/Scenes/HomeView/useHomeViewTracking"
+import { useHomeViewTracking } from "app/Scenes/HomeView/hooks/useHomeViewTracking"
 import { navigate } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
 import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
